requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
python-dotenv==1.0.1
pydantic-settings==2.3.4
line-bot-sdk==3.17.1
gspread==6.1.0
oauth2client==4.1.2